# cloud/internal/httpapi/worker_orchestration_handlers.go
package httpapi

import (
	"context"
	"net/http"
	"strings"

	"github.com/aoagents/agent-orchestrator/cloud/internal/domain"
	"github.com/aoagents/agent-orchestrator/cloud/internal/worker"
	"github.com/go-chi/chi/v5"
)

// orchestratorProviderStore reads the sandbox provider a parent orchestrator
// runs on, so a child it spawns inherits it rather than the control plane
// default. It mirrors the workerCredentialAvailabilityStore narrow-interface
// pattern so the concrete store carries the method without widening Store.
type orchestratorProviderStore interface {
	OrchestratorSandboxProvider(context.Context, string, string) (string, error)
}

// orchestratorWorkerAgentStore resolves the worker agent the parent
// orchestrator's project was configured with (config.worker.agent), so a child
// spawned without an explicit harness inherits exactly that instead of a
// hardcoded default. Same narrow-interface pattern as above.
type orchestratorWorkerAgentStore interface {
	OrchestratorProjectWorkerAgent(context.Context, string, string) (string, error)
}

type createWorkerChildRequest struct {
	Harness                     string   `json:"harness"`
	DisplayName                 string   `json:"displayName"`
	Prompt                      string   `json:"prompt"`
	Mode                        string   `json:"mode,omitempty"`
	DeniedCommands              []string `json:"deniedCommands,omitempty"`
	SandboxProviderConnectionID string   `json:"sandboxProviderConnectionId,omitempty"`
}

func (s *Server) listWorkerChildren(w http.ResponseWriter, r *http.Request) {
	claims, ok := requireOrchestratorScope(w, r)
	if !ok {
		return
	}
	limit, err := parseLimit(r)
	if err != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_request", err.Error())
		return
	}
	cursor, err := parseCursor(r.URL.Query().Get("cursor"))
	if err != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_cursor", "The pagination cursor is invalid.")
		return
	}
	// Terminated children are noise for day-to-day orchestration, so the
	// default view hides them; `ao list --all` opts back in for history.
	includeTerminated := r.URL.Query().Get("includeTerminated") == "true"
	children, hasMore, err := s.store.ListOrchestratorChildren(
		r.Context(), claims.OrgID, claims.SessionID, includeTerminated, cursor, limit,
	)
	if err != nil {
		s.writeStoreError(w, r, err)
		return
	}
	items, err := s.childItems(r, claims.OrgID, children)
	if err != nil {
		s.writeStoreError(w, r, err)
		return
	}
	page := pageInfo{HasMore: hasMore}
	if hasMore && len(children) > 0 {
		last := children[len(children)-1]
		page.NextCursor = encodeCursor(last.UpdatedAt, last.ID)
	}
	writeJSON(w, http.StatusOK, map[string]any{"items": items, "page": page})
}

// childItems renders child sessions with their PR facts (for status
// derivation) and full PR rows (for the prs[] payload) in two batch reads.
func (s *Server) childItems(
	r *http.Request,
	orgID string,
	children []domain.Session,
) ([]sessionChildResponse, error) {
	childIDs := make([]string, len(children))
	for i, child := range children {
		childIDs[i] = child.ID
	}
	prFacts, err := s.store.PRFactsBySession(r.Context(), orgID, childIDs)
	if err != nil {
		return nil, err
	}
	pullRequests, err := s.store.PullRequestsBySessions(r.Context(), orgID, childIDs)
	if err != nil {
		return nil, err
	}
	items := make([]sessionChildResponse, 0, len(children))
	for _, child := range children {
		items = append(items, toSessionChildResponse(child, prFacts[child.ID], pullRequests[child.ID]))
	}
	return items, nil
}

func (s *Server) createWorkerChild(w http.ResponseWriter, r *http.Request) {
	claims, ok := requireOrchestratorScope(w, r)
	if !ok {
		return
	}
	key, err := idempotencyKey(r)
	if err != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_request", err.Error())
		return
	}
	var request createWorkerChildRequest
	if err := decodeJSON(w, r, &request); err != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_request", "The request body is invalid.")
		return
	}
	request.Harness = strings.TrimSpace(request.Harness)
	request.DisplayName = strings.TrimSpace(request.DisplayName)
	request.Prompt = strings.TrimSpace(request.Prompt)
	request.Mode = strings.TrimSpace(request.Mode)
	request.SandboxProviderConnectionID = strings.TrimSpace(request.SandboxProviderConnectionID)
	if request.Mode == "" {
		request.Mode = "trusted"
	}
	// The project's configured worker agent (config.worker.agent) is authoritative
	// for orchestrator-spawned workers: the harness must match what was chosen at
	// project creation regardless of what the orchestrator asks for. Some agents
	// (e.g. Codex) spawn children naming their own harness, which would otherwise
	// override the project's choice. Resolve and force it here, before the
	// credential check and the provisioning plan. Only when the project set no
	// worker agent do we honor the requested harness, falling back to claude-code.
	if workerAgentStore, ok := s.store.(orchestratorWorkerAgentStore); ok {
		configured, err := workerAgentStore.OrchestratorProjectWorkerAgent(r.Context(), claims.OrgID, claims.SessionID)
		if err != nil {
			s.writeStoreError(w, r, err)
			return
		}
		if configured = strings.TrimSpace(configured); configured != "" {
			request.Harness = configured
		}
	}
	if request.Harness == "" {
		request.Harness = "claude-code"
	}
	if request.Prompt == "" {
		writeError(w, r, http.StatusUnprocessableEntity, "validation_error", "Child prompt is required.")
		return
	}
	validation := createSessionRequest{
		ProjectID: claims.SessionID, Kind: "worker",
		Harness: request.Harness, DisplayName: request.DisplayName,
		Prompt: request.Prompt, Mode: request.Mode,
		DeniedCommands:              request.DeniedCommands,
		SandboxProviderConnectionID: request.SandboxProviderConnectionID,
	}
	if !validSessionInput(validation) {
		writeError(w, r, http.StatusUnprocessableEntity, "validation_error", "Child session input is invalid.")
		return
	}
	if !supportedInteractivePolicy(validation) {
		writeError(
			w, r, http.StatusUnprocessableEntity, "unsupported_policy",
			"Interactive child agents support standard or trusted mode without command deny rules.",
		)
		return
	}
	credentialStore, ok := s.store.(workerCredentialAvailabilityStore)
	if !ok {
		writeError(
			w, r, http.StatusNotImplemented, "not_implemented",
			"Agent provider checks are unavailable.",
		)
		return
	}
	available, err := credentialStore.OrchestratorAgentCredentialAvailable(
		r.Context(), claims.OrgID, claims.SessionID, request.Harness,
	)
	if err != nil {
		s.writeStoreError(w, r, err)
		return
	}
	if !available {
		writeError(
			w, r, http.StatusUnprocessableEntity,
			"agent_provider_required",
			"Connect and validate the selected coding-agent provider before spawning a child.",
		)
		return
	}
	providerStore, ok := s.store.(orchestratorProviderStore)
	if !ok {
		writeError(
			w, r, http.StatusNotImplemented, "not_implemented",
			"Sandbox provider inheritance is unavailable.",
		)
		return
	}
	parentProvider, err := providerStore.OrchestratorSandboxProvider(
		r.Context(), claims.OrgID, claims.SessionID,
	)
	if err != nil {
		s.writeStoreError(w, r, err)
		return
	}
	// A child always runs on its orchestrator's provider, never the control
	// plane default: a NodeOps orchestrator spawns NodeOps workers and a Coder
	// orchestrator spawns Coder workers, even on a control plane that offers
	// both. The provider is read from the parent's immutable sandbox row, so it
	// is unaffected by the client-side provider toggle (which only stamps the
	// provider for top-level sessions the app creates).
	plan, err := s.provisioning.SessionPlanForProvider(request.Harness, parentProvider)
	if err != nil {
		s.logger.Error("resolve child sandbox provisioning plan", "error", err, "request_id", requestID(r))
		writeError(w, r, http.StatusInternalServerError, "internal_error", "Sandbox provisioning is misconfigured.")
		return
	}
	child, err := s.store.CreateOrchestratorChild(
		r.Context(), claims.OrgID, claims.SessionID, key, s.maxSandboxes,
		domain.CreateSession{
			Kind: "worker", Harness: request.Harness, DisplayName: request.DisplayName,
			Prompt: request.Prompt, Mode: request.Mode, DeniedCommands: request.DeniedCommands,
			Provider: plan.Provider, SandboxConnectionID: request.SandboxProviderConnectionID,
			ResourceProfile: plan.ResourceProfile, BootstrapContext: plan.BootstrapContext,
			Release: s.release,
		},
	)
	if err != nil {
		s.writeStoreError(w, r, err)
		return
	}
	writeJSON(w, http.StatusCreated, map[string]any{"session": toSessionResponse(child, nil)})
}

func (s *Server) sendWorkerChildMessage(w http.ResponseWriter, r *http.Request) {
	claims, ok := requireOrchestratorScope(w, r)
	if !ok {
		return
	}
	childID := chi.URLParam(r, "sessionId")
	if requireUUID(childID, "sessionId") != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_request", "sessionId must be a UUID.")
		return
	}
	key, err := idempotencyKey(r)
	if err != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_request", err.Error())
		return
	}
	var request sendMessageRequest
	if err := decodeJSON(w, r, &request); err != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_request", "The request body is invalid.")
		return
	}
	if validateSendMessageRequest(request) != nil {
		writeError(w, r, http.StatusUnprocessableEntity, "validation_error", "Message text must be between 1 and 65536 bytes.")
		return
	}
	var event domain.ClientEvent
	if len(request.AttachmentIDs) > 0 {
		store, ok := s.store.(interface {
			SendOrchestratorChildMessageWithAttachments(context.Context, string, string, string, string, string, []string) (domain.ClientEvent, error)
		})
		if !ok {
			writeError(w, r, 503, "attachments_unavailable", "Image forwarding is not available.")
			return
		}
		event, err = store.SendOrchestratorChildMessageWithAttachments(r.Context(), claims.OrgID, claims.SessionID, childID, key, request.Text, request.AttachmentIDs)
	} else {
		event, err = s.store.SendOrchestratorChildMessage(r.Context(), claims.OrgID, claims.SessionID, childID, key, request.Text)
	}
	if err != nil {
		s.writeStoreError(w, r, err)
		return
	}
	writeJSON(w, http.StatusAccepted, map[string]any{"event": toClientEventResponse(event)})
}

func (s *Server) deleteWorkerChild(w http.ResponseWriter, r *http.Request) {
	claims, ok := requireOrchestratorScope(w, r)
	if !ok {
		return
	}
	childID := chi.URLParam(r, "sessionId")
	if requireUUID(childID, "sessionId") != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_request", "sessionId must be a UUID.")
		return
	}
	if err := s.store.DeleteOrchestratorChild(
		r.Context(), claims.OrgID, claims.SessionID, childID,
	); err != nil {
		s.writeStoreError(w, r, err)
		return
	}
	writeJSON(w, http.StatusAccepted, map[string]any{
		"session": map[string]any{
			"id": childID, "desiredState": domain.SandboxDesiredDeleted,
		},
	})
}

// reportToParent delivers a child worker's message into the conversation of
// the orchestrator that spawned it. The worker:report scope is issued only to
// sessions with a parent, and the store re-verifies the parent link, so a
// worker can never message anyone but its own orchestrator.
func (s *Server) reportToParent(w http.ResponseWriter, r *http.Request) {
	claims := workerFrom(r)
	if !worker.HasScope(claims, "worker:report") {
		writeError(w, r, http.StatusForbidden, "SCOPE_REQUIRED", "The worker:report scope is required.")
		return
	}
	key, err := idempotencyKey(r)
	if err != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_request", err.Error())
		return
	}
	var request sendMessageRequest
	if err := decodeJSON(w, r, &request); err != nil {
		writeError(w, r, http.StatusBadRequest, "invalid_request", "The request body is invalid.")
		return
	}
	if validateSendMessageRequest(request) != nil {
		writeError(w, r, http.StatusUnprocessableEntity, "validation_error", "Message text must be between 1 and 65536 bytes.")
		return
	}
	var event domain.ClientEvent
	if len(request.AttachmentIDs) > 0 {
		store, ok := s.store.(interface {
			ReportToOrchestratorWithAttachments(context.Context, string, string, string, string, []string) (domain.ClientEvent, error)
		})
		if !ok {
			writeError(w, r, 503, "attachments_unavailable", "Image forwarding is not available.")
			return
		}
		event, err = store.ReportToOrchestratorWithAttachments(r.Context(), claims.OrgID, claims.SessionID, key, request.Text, request.AttachmentIDs)
	} else {
		event, err = s.store.ReportToOrchestrator(r.Context(), claims.OrgID, claims.SessionID, key, request.Text)
	}
	if err != nil {
		s.writeStoreError(w, r, err)
		return
	}
	writeJSON(w, http.StatusAccepted, map[string]any{"event": toClientEventResponse(event)})
}

func requireOrchestratorScope(w http.ResponseWriter, r *http.Request) (worker.Claims, bool) {
	claims := workerFrom(r)
	if !worker.HasScope(claims, "worker:orchestrate") {
		writeError(w, r, http.StatusForbidden, "SCOPE_REQUIRED", "The worker:orchestrate scope is required.")
		return worker.Claims{}, false
	}
	return claims, true
}