# contracts/cloud/openapi.yaml
openapi: 3.1.0
info:
  title: AO Cloud Public API
  version: 1.0.0
  description: |
    Client-facing contract for a separate AO Cloud implementation. This
    repository does not serve these routes. The worker surface documents the
    exact outbound HTTP protocol used by sandbox workers, including one-time
    bootstrap, rotating credentials, fenced turn execution, orchestration, and
    durable workspace and terminal transport.
security:
  - bearerAuth: []
...
paths:
...
  /api/cloud/v1/orgs/{orgId}/sessions:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    get:
      operationId: listSessions
      tags: [Sessions]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: A page of sessions in the tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createSession
      tags: [Sessions]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionInput"
      responses:
        "201":
          description: Session created.
          content:
            application/json:
              schema:
                type: object
                required: [session]
                properties:
                  session:
                    $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
...
  /api/cloud/v1/worker/children:
    get:
      operationId: listWorkerChildren
      tags: [Worker Orchestration]
      security:
        - workerAuth: []
      description: >-
        Lists this orchestrator's direct child sessions with their pull
        requests. Terminated children are hidden unless includeTerminated.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: includeTerminated
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: A page of direct child sessions owned by this orchestrator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionChildrenPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/WorkerUnauthorized"
        "403":
          $ref: "#/components/responses/WorkerScopeRequired"
        "404":
          $ref: "#/components/responses/Error"
    post:
      operationId: createWorkerChild
      tags: [Worker Orchestration]
      security:
        - workerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkerChildInput"
      responses:
        "201":
          description: A worker child was created under this orchestrator.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [session]
                properties:
                  session:
                    $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/WorkerUnauthorized"
        "403":
          $ref: "#/components/responses/WorkerScopeRequired"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/Error"
  /api/cloud/v1/worker/children/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    delete:
      operationId: deleteWorkerChild
      tags: [Worker Orchestration]
      security:
        - workerAuth: []
      description: Requests deletion of a direct child owned by the authenticated orchestrator.
      responses:
        "202":
          description: Child sandbox deletion was requested.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSessionResponse"
        "401":
          $ref: "#/components/responses/WorkerUnauthorized"
        "403":
          $ref: "#/components/responses/WorkerScopeRequired"
        "404":
          $ref: "#/components/responses/Error"
  /api/cloud/v1/worker/children/{sessionId}/messages:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      operationId: sendWorkerChildMessage
      tags: [Worker Orchestration]
      security:
        - workerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "202":
          description: The message was durably appended to the direct child.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [event]
                properties:
                  event:
                    $ref: "#/components/schemas/UserMessageEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/WorkerUnauthorized"
        "403":
          $ref: "#/components/responses/WorkerScopeRequired"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/cloud/v1/worker/parent/messages:
    post:
      operationId: reportToParent
      tags: [Worker Orchestration]
      security:
        - workerAuth: []
      description: >-
        Delivers a child worker's message into the conversation of the
        orchestrator that spawned it. Requires the worker:report scope, which
        is issued only to sessions with an orchestrator parent.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "202":
          description: The message was durably appended to the parent orchestrator.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [event]
                properties:
                  event:
                    $ref: "#/components/schemas/UserMessageEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/WorkerUnauthorized"
        "403":
          $ref: "#/components/responses/WorkerScopeRequired"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    workerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Session-scoped credential formatted as `Worker <token>`. Every
        authenticated route verifies both the token signature and that its
        worker ID and epoch are still current. Replacing a sandbox therefore
        fences an otherwise unexpired token with `STALE_WORKER_TOKEN`.

        Route scopes are `worker:connect` for heartbeat, `worker:event` for
        events, `worker:turn:claim`, `worker:turn:poll`, and
        `worker:turn:complete` for turn execution,
        `worker:credential:read` for the agent secret, `worker:git` for a
        checkout grant, `worker:orchestrate` for child controls, and
        `worker:transport` for workspace and terminal transport, and
        `worker:notification` for durable agent notification intake.
  parameters:
...
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Reusing a key with the same command returns the original result.
        Reusing it with a different command returns an IDEMPOTENCY_CONFLICT.
      schema:
        type: string
        minLength: 1
        maxLength: 200
  responses:
    Error:
      description: Standard Cloud API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequest:
      description: The JSON body, path value, query value, or required header is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ValidationError:
      description: The request is syntactically valid but violates a resource constraint.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    WorkerUnauthorized:
      description: |
        The `Worker` credential is absent, invalid, expired, or belongs to a
        worker epoch that has been replaced.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    WorkerScopeRequired:
      description: The worker token does not contain the scope required by this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    AgentCapability:
      type: string
      enum:
        - interface.chat
        - interface.tui
        - model.catalog
        - model.custom
        - attachments
        - browser.preview
        - review.execute
        - session.resume
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, code, message, requestId]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
    PageInfo:
      type: object
      additionalProperties: false
      required: [hasMore]
      properties:
        hasMore:
          type: boolean
        nextCursor:
          type: string
...
    SessionKind:
      type: string
      enum: [worker, orchestrator]
    SessionMode:
      type: string
      enum: [read-only, standard, trusted]
    SessionActivityState:
      type: string
      enum: [active, idle, waiting_input, blocked, exited]
    SessionStatus:
      type: string
      enum:
        - working
        - needs_input
        - pr_open
        - draft
        - review_pending
        - ci_failed
        - changes_requested
        - approved
        - mergeable
        - merged
        - exited
        - idle
        - terminated
        - no_signal
    Turn:
      type: object
      additionalProperties: false
      required: [id, sessionId, userMessageSequence, state, attemptCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        sessionId:
          type: string
        userMessageSequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        state:
          type: string
          enum: [queued, provisioning, running, cancel_requested, completed, failed]
        attemptCount:
          type: integer
          minimum: 0
        errorMessage:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateWorkerChildInput:
      type: object
      additionalProperties: false
      required: [harness, displayName, prompt]
      properties:
        harness:
          type: string
          minLength: 1
          maxLength: 120
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        prompt:
          type: string
          minLength: 1
          maxLength: 65536
        mode:
          $ref: "#/components/schemas/SessionMode"
        deniedCommands:
          type: array
          maxItems: 128
          items:
            type: string
            minLength: 1
            maxLength: 512
        sandboxProviderConnectionId:
          type: string
          format: uuid
...
    SendMessageInput:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        attachmentIds:
          $ref: "#/components/schemas/AttachmentIds"
        text:
          type: string
          maxLength: 65536
...
    DeleteSessionResponse:
      type: object
      additionalProperties: false
      required: [session]
      properties:
        session:
          type: object
          additionalProperties: false
          required: [id, desiredState]
          properties:
            id:
              type: string
              format: uuid
            desiredState:
              type: string
              const: deleted
    SessionInterfaceMode:
      type: string
      enum: [tui, chat]
...
    Session:
      type: object
      additionalProperties: false
      required:
        - id
        - orgId
        - projectId
        - kind
        - harness
        - displayName
        - branch
        - mode
        - deniedCommands
        - interfaceMode
        - activityState
        - status
        - runtimeConnected
        - isTerminated
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/SessionKind"
        harness:
          type: string
        displayName:
          type: string
        branch:
          type: string
        mode:
          $ref: "#/components/schemas/SessionMode"
        model:
          type: string
          description: >-
            Coding-agent model the session launched with; empty means the harness
            default.
        deniedCommands:
          type: array
          items:
            type: string
        interfaceMode:
          $ref: "#/components/schemas/SessionInterfaceMode"
        activityState:
          $ref: "#/components/schemas/SessionActivityState"
        status:
          $ref: "#/components/schemas/SessionStatus"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/AgentCapability"
        runtimeConnected:
          type: boolean
        runtimeState:
          type: string
        runtimeError:
          type: string
        activeTurn:
          $ref: "#/components/schemas/Turn"
        isTerminated:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSessionInput:
      type: object
      additionalProperties: false
      required: [projectId, kind, harness, displayName, prompt]
      properties:
        attachmentIds:
          $ref: "#/components/schemas/AttachmentIds"
        projectId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/SessionKind"
        harness:
          type: string
          minLength: 1
          maxLength: 120
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        prompt:
          type: string
          maxLength: 65536
        mode:
          $ref: "#/components/schemas/SessionMode"
          default: trusted
        model:
          type: string
          maxLength: 200
          description: >-
            Coding-agent model the session launches with (harness-native id, e.g.
            "anthropic/claude-opus-4-8"). Optional; empty uses the harness default.
        deniedCommands:
          type: array
          items:
            type: string
          default: []
        sandboxProviderConnectionId:
          type: string
          format: uuid
    SessionPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Session"
        page:
          $ref: "#/components/schemas/PageInfo"
    SessionPullRequestFacts:
      type: object
      additionalProperties: false
      required: [url, number, state, ci, review, mergeability, reviewComments, updatedAt]
      properties:
        url:
          type: string
        number:
          type: integer
        state:
          $ref: "#/components/schemas/PullRequestState"
        ci:
          $ref: "#/components/schemas/CIState"
        review:
          $ref: "#/components/schemas/ReviewDecision"
        mergeability:
          $ref: "#/components/schemas/MergeabilityState"
        reviewComments:
          type: boolean
        sourceBranch:
          type: string
        targetBranch:
          type: string
        updatedAt:
          type: string
          format: date-time
    SessionWithPullRequests:
      allOf:
        - $ref: "#/components/schemas/Session"
        - type: object
          required: [prs]
          properties:
            prs:
              type: array
              items:
                $ref: "#/components/schemas/SessionPullRequestFacts"
    SessionChildrenPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SessionWithPullRequests"
        page:
          $ref: "#/components/schemas/PageInfo"
    PullRequestState:
      type: string
      enum: [draft, open, merged, closed]
    CIState:
      type: string
      enum: [unknown, pending, passing, failing]
...
    ReviewDecision:
      type: string
      enum: [none, approved, changes_requested, review_required]
    MergeabilityState:
      type: string
      enum: [unknown, mergeable, conflicting, blocked, unstable]
...
    UserMessageEvent:
      type: object
      additionalProperties: false
      required: [sessionId, sequence, type, payload, createdAt]
      properties:
        sessionId:
          type: string
        sequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        type:
          const: chat.user_message
        payload:
          type: object
          additionalProperties: false
          required: [text]
          properties:
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/ImageAttachment"
            text:
              type: string
            turnId:
              type: string
              format: uuid
        createdAt:
          type: string
          format: date-time
...
    AttachmentIds:
      type: array
      maxItems: 8
      uniqueItems: true
      items:
        type: string
        format: uuid
    ImageAttachment:
      type: object
      required: [id, filename, size, mimeType, sha256, status]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
          maxLength: 255
        size:
          type: integer
          minimum: 1
          maximum: 10485760
        mimeType:
          type: string
          enum: [image/png, image/jpeg, image/webp, image/gif, image/bmp]
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
        status:
          type: string
          enum: [pending, ready, expired]
...